Update Simulator.py
Filled in more of neutron travel code, altering return values to allow for neutron tracking and messages for highest level of output verbosity. (ReactorSim.DoNeutronTravel() method)

Added new level of messaging verbosity: 'ABSURD.'  This is the only level at which neutron location information will be added to what is returned by the simulator.